# specs/001-blog-post-crud/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Blog Post CRUD API
  description: 블로그 포스트 생성, 조회, 수정, 삭제 API
  version: 1.0.0
  contact:
    name: Portfolio Blog API

servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production

tags:
  - name: posts
    description: 블로그 포스트 관리
  - name: tags
    description: 태그 관리

paths:
  /posts:
    get:
      tags:
        - posts
      summary: 포스트 목록 조회
      description: 공개된 블로그 포스트 목록을 페이지네이션으로 조회합니다.
      operationId: listPosts
      parameters:
        - name: page
          in: query
          description: 페이지 번호 (1부터 시작)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: 페이지당 포스트 수
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: status
          in: query
          description: 포스트 상태 필터 (관리자 전용)
          schema:
            $ref: '#/components/schemas/PostStatus'
        - name: tag
          in: query
          description: 태그로 필터링
          schema:
            type: string
      responses:
        '200':
          description: 포스트 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'

    post:
      tags:
        - posts
      summary: 포스트 생성
      description: 새 블로그 포스트를 생성합니다. 관리자 인증 필요.
      operationId: createPost
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '201':
          description: 포스트 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  /posts/{slug}:
    get:
      tags:
        - posts
      summary: 포스트 상세 조회
      description: Slug로 특정 포스트를 조회합니다.
      operationId: getPost
      parameters:
        - name: slug
          in: path
          required: true
          description: 포스트 URL 식별자
          schema:
            type: string
            example: my-first-blog-post
      responses:
        '200':
          description: 포스트 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - posts
      summary: 포스트 수정
      description: 기존 포스트를 수정합니다. 관리자 인증 필요.
      operationId: updatePost
      security:
        - ApiKeyAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          description: 포스트 URL 식별자
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: 포스트 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - posts
      summary: 포스트 삭제
      description: 포스트를 영구 삭제합니다. 관리자 인증 필요.
      operationId: deletePost
      security:
        - ApiKeyAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          description: 포스트 URL 식별자
          schema:
            type: string
      responses:
        '204':
          description: 포스트 삭제 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /tags:
    get:
      tags:
        - tags
      summary: 태그 목록 조회
      description: 모든 태그 목록을 조회합니다.
      operationId: listTags
      responses:
        '200':
          description: 태그 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagResponse'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 관리자 API 키 (임시, 추후 JWT로 교체 예정)

  schemas:
    PostStatus:
      type: string
      enum:
        - draft
        - published
      description: 포스트 상태

    PostCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: 포스트 제목
          example: 나의 첫 번째 블로그 포스트
        content:
          type: string
          maxLength: 50000
          description: 포스트 본문 (Markdown)
          example: "# Hello World\n\n이것은 첫 번째 포스트입니다."
        status:
          $ref: '#/components/schemas/PostStatus'
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          description: 태그 목록
          example: ["Python", "FastAPI"]

    PostUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: 포스트 제목
        content:
          type: string
          maxLength: 50000
          description: 포스트 본문 (Markdown)
        status:
          $ref: '#/components/schemas/PostStatus'
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          description: 태그 목록

    PostResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 포스트 고유 ID
        title:
          type: string
          description: 포스트 제목
        slug:
          type: string
          description: URL 식별자
        content:
          type: string
          description: 포스트 본문
        status:
          $ref: '#/components/schemas/PostStatus'
        created_at:
          type: string
          format: date-time
          description: 생성 일시
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: 수정 일시
        published_at:
          type: string
          format: date-time
          nullable: true
          description: 최초 공개 일시
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'

    PostListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        status:
          $ref: '#/components/schemas/PostStatus'
        created_at:
          type: string
          format: date-time
        excerpt:
          type: string
          nullable: true
          description: 본문 미리보기 (처음 200자)
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'

    PostListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PostListItem'
        total:
          type: integer
          description: 전체 포스트 수
        page:
          type: integer
          description: 현재 페이지
        page_size:
          type: integer
          description: 페이지당 포스트 수
        total_pages:
          type: integer
          description: 전체 페이지 수

    TagResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: 태그 이름
        slug:
          type: string
          description: 태그 URL 식별자

    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: 에러 타입 URI
        title:
          type: string
          description: 에러 제목
        status:
          type: integer
          description: HTTP 상태 코드
        detail:
          type: string
          description: 상세 에러 메시지
        instance:
          type: string
          description: 요청 경로

  responses:
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: "https://example.com/errors/not-found"
            title: "Not Found"
            status: 404
            detail: "Post not found"
            instance: "/api/v1/posts/non-existent-slug"

    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: "https://example.com/errors/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Invalid or missing API key"
            instance: "/api/v1/posts"

    ValidationError:
      description: 유효성 검사 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: "https://example.com/errors/validation"
            title: "Validation Error"
            status: 422
            detail: "Title is required"
            instance: "/api/v1/posts"